requirements.txt
# arxiv-rag — root requirements
# Install with: pip install -r requirements.txt
#
# For the sparse embedder service, use sparse_embedder/requirements.txt
# For the RAG proxy API service,   use api/requirements.txt

# ── Vector store ──────────────────────────────────────────────────────────────
qdrant-client>=1.10.0

# ── Embeddings and reranking ──────────────────────────────────────────────────
FlagEmbedding>=1.2.11          # BGE-M3 dense + sparse + reranker
sentence-transformers>=2.7.0   # fallback / utility

# ── Orchestration ─────────────────────────────────────────────────────────────
langgraph>=0.1.0
langchain-core>=0.2.0

# ── Observability ─────────────────────────────────────────────────────────────
langfuse>=2.0.0

# ── HTTP clients ──────────────────────────────────────────────────────────────
httpx>=0.27.0
requests>=2.31.0

# ── Caching ───────────────────────────────────────────────────────────────────
redis>=5.0.0

# ── PDF ingestion (05_ingest_pdfs.py, 06_caption_figures.py) ─────────────────
PyMuPDF>=1.24.0
unstructured[pdf]>=0.14.0
tiktoken>=0.7.0

# ── Clustering / classification (05_ingest_pdfs.py) ──────────────────────────
hdbscan>=0.8.33
umap-learn>=0.5.6
scikit-learn>=1.4.0

# ── Dataset download (01_download_arxiv.py) ───────────────────────────────────
datasets>=2.18.0
huggingface_hub>=0.22.0

# ── Numerical ─────────────────────────────────────────────────────────────────
numpy>=1.26.0

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm>=4.66.0
pydantic>=2.0.0
